respond to PR comments
