+++ workflows/prognostic_c48_run/runtime/transformers/tendency_prescriber.py
@@ -10,7 +10,7 @@ import loaders
 from runtime.monitor import Monitor
 from runtime.types import Diagnostics, Step, State
 from runtime.derived_state import DerivedFV3State
-from runtime.conversions import quantity_state_to_dataset, dataset_to_quantity_state
+from runtime.scatter import scatter_within_tile
 
 logger = logging.getLogger(__name__)
 
@@ -104,33 +104,3 @@ class TendencyPrescriber:
 
         step.__name__ = func.__name__
         return step
-
-
-def scatter_within_tile(
-    time: cftime.DatetimeJulian,
-    time_lookup_function: Callable[[cftime.DatetimeJulian], State],
-    communicator: pace.util.CubedSphereCommunicator,
-) -> xr.Dataset:
-    """Scatter data for a timestamp from each tile's master rank to its subranks.
-
-    Args:
-        time: time of the data
-        time_lookup_function: a function that takes a time and returns a state dict
-            containing data arrays to be scattered
-        communicator: model cubed sphere communicator
-
-    Returns:
-        Dataset of scattered data arrays
-    """
-    if communicator.tile.rank == 0:
-        state: State = time_lookup_function(time)
-    else:
-        state = {}
-
-    tile = communicator.partitioner.tile_index(communicator.rank)
-    if communicator.tile.rank == 0:
-        ds = xr.Dataset(state).isel(tile=tile)
-    else:
-        ds = xr.Dataset()
-    tendencies = communicator.tile.scatter_state(dataset_to_quantity_state(ds))
-    return quantity_state_to_dataset(tendencies)
